backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
msal==1.25.0
requests==2.31.0
beautifulsoup4==4.12.2
cohere==5.5.0
python-dotenv==1.0.0
gunicorn==21.2.0